виніс поп ап у окремий файл, але не закривається
